% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carte_banqDens.R
\name{banq.dens}
\alias{banq.dens}
\title{Carte affichant la densité des banques pour une zone d'emploi}
\usage{
banq.dens(ze, type, bdd_zese, bdd_coordonnees_banques2022)
}
\arguments{
\item{ze}{character}

\item{type}{character (Lucrative/Coopérative/Tout)}

\item{bdd_zese}{data frame}

\item{bdd_coordonnees_banques2022}{data frame}
}
\value{
carte
}
\description{
Carte affichant la densité des banques pour une zone d'emploi
}
