% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bdd_coordonnees_ze_banques2022}
\alias{bdd_coordonnees_ze_banques2022}
\title{Données des coordonnées des banques en 2022 avec leur zone d'emploi}
\format{
Une data frame avec six variables :
\describe{
\item{\code{Banque}}{Nom de la banque}
\item{\code{Type}}{Type de la banque}
\item{\code{Adresse}}{Adresse de la banque}
\item{\code{Longitude}}{Longitude de la banque}
\item{\code{Latitude}}{Latitude de la banque}
\item{\code{ze}}{Zone d'emploi de la banque}
}
}
\usage{
bdd_coordonnees_ze_banques2022
}
\description{
Données des coordonnées des banques en 2022 avec leur zone d'emploi
}
\keyword{datasets}
