% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bdd_effectif_banques2022_ze2020}
\alias{bdd_effectif_banques2022_ze2020}
\title{Données du nombre des banques en 2022 en fonction de leur zone d'emploi}
\format{
Une data frame avec six variables :
}
\usage{
bdd_effectif_banques2022_ze2020
}
\description{
Données du nombre des banques en 2022 en fonction de leur zone d'emploi
}
\keyword{datasets}
