% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bdd_social_ze2020}
\alias{bdd_social_ze2020}
\title{Données de critères socio-économique}
\format{
Une data frame avec 34 variables
}
\usage{
bdd_social_ze2020
}
\description{
Les données sont isues des années entre 2017 à 2020
}
\keyword{datasets}
