% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bdd_zese}
\alias{bdd_zese}
\title{Données de critères socio-économique avec geometry}
\format{
Une data frame avec 35 variables:
}
\usage{
bdd_zese
}
\description{
Les données sont isues des années entre 2017 à 2020
}
\keyword{datasets}
