% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarteV4.R
\name{gg_map_ze}
\alias{gg_map_ze}
\title{Version 4 de la carte en ggplot qui afficher uniquement par zone d'emploi}
\usage{
gg_map_ze(bdd_zese, bdd_coordonnees_ze_banques2022, type, ze)
}
\arguments{
\item{bdd_zese}{data frame}

\item{bdd_coordonnees_ze_banques2022}{data frame}

\item{type}{character}

\item{ze}{character}
}
\value{
carte ggplot
}
\description{
Version 4 de la carte en ggplot qui afficher uniquement par zone d'emploi
}
