% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarteV3.R
\name{plotly_map}
\alias{plotly_map}
\title{Version 3 de la carte en plotly}
\usage{
plotly_map(bdd_coordonnees_banques2022)
}
\arguments{
\item{bdd_coordonnees_banques2022}{data frame}
}
\value{
carte plotly
}
\description{
Version 3 de la carte en plotly
}
