% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recup_long_lat.R
\name{recup.longlat}
\alias{recup.longlat}
\title{Fonction qui permet de récupérer la longitude et la latitude des frontières d'une zone d'emploi}
\usage{
recup.longlat(indice, bdd_zese)
}
\arguments{
\item{indice}{integer}

\item{bdd_zese}{data frame bdd_zese}
}
\value{
data frame
}
\description{
Fonction qui permet de récupérer la longitude et la latitude des frontières d'une zone d'emploi
}
