% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarteV2.R
\name{gg_map}
\alias{gg_map}
\title{Version 2 de la carte en ggplot}
\usage{
gg_map(bdd_zese, crit, label, point, data)
}
\arguments{
\item{bdd_zese}{data frame}

\item{crit}{integer}

\item{label}{character}

\item{point}{interger}

\item{data}{data frame}
}
\value{
carte ggplot
}
\description{
Version 2 de la carte en ggplot
}
